// src/app/(sistema)/(planilla)/personal/components/types/personal.ts
export type PersonalGeneral = {
    documento: string
    n_documento: string
    fecha_nacimiento: string
    genero: string
    celular: string
    telefono: string
    correo: string
    direccion: string
    nivel_educativo: string
    carrera_profesional: string
    estado_civil: string
    licencia: string
}

export type PersonalLaboral = {
    area: string
    cargo: string
    tipo_trabajador: string
    sede: string
    turno: string
    salario: string
    fecha_vinculacion: string
    fecha_retiro: string
    forma_pago: string
    banco: string
    numero_cuenta: string
    seguro_salud: string
    tipo_contrato: string
    regimen: string
    estado: string
}